+++ examples/stub_domains_private/outputs.tf
@@ -30,5 +30,5 @@ output "ca_certificate" {
 
 output "service_account" {
   description = "The service account to default running nodes as if not overridden in `node_pools`."
-  value       = "${local.service_account}"
+  value       = "${module.gke.service_account}"
 }