+++ test/fixtures/stub_domains_private/main.tf
@@ -14,27 +14,26 @@
  * limitations under the License.
  */
 
-
 resource "random_string" "suffix" {
   length  = 4
   special = false
   upper   = false
 }
 
-provider "google" {
-  project     = "${var.project_id}"
-}
-
 resource "google_compute_network" "main" {
-  name                    = "cft-gke-test-${random_string.suffix.result}"
+  name = "cft-gke-test-${random_string.suffix.result}"
+
   auto_create_subnetworks = "false"
+  project                 = "${var.project_id}"
 }
 
 resource "google_compute_subnetwork" "main" {
-  name          = "cft-gke-test-${random_string.suffix.result}"
   ip_cidr_range = "10.0.0.0/17"
-  region        = "${var.region}"
+  name          = "cft-gke-test-${random_string.suffix.result}"
   network       = "${google_compute_network.main.self_link}"
+
+  project = "${var.project_id}"
+  region  = "${var.region}"
 
   secondary_ip_range {
     range_name    = "cft-gke-test-pods-${random_string.suffix.result}"
@@ -45,4 +44,17 @@ resource "google_compute_subnetwork" "main" {
     range_name    = "cft-gke-test-services-${random_string.suffix.result}"
     ip_cidr_range = "192.168.64.0/18"
   }
+}
+
+module "example" {
+  source = "../../../examples/stub_domains_private"
+
+  compute_engine_service_account = "${var.compute_engine_service_account}"
+  ip_range_pods                  = "${google_compute_subnetwork.main.secondary_ip_range.0.range_name}"
+  ip_range_services              = "${google_compute_subnetwork.main.secondary_ip_range.1.range_name}"
+  network                        = "${google_compute_network.main.name}"
+  project_id                     = "${var.project_id}"
+  cluster_name_suffix            = "-${random_string.suffix.result}"
+  region                         = "${var.region}"
+  subnetwork                     = "${google_compute_subnetwork.main.name}"
 }